+++ server/clic/subset.py
@@ -6,6 +6,7 @@ Returns subsets of given texts, for example quotations.
 - corpora: 1+ corpus name (e.g. 'dickens') or book name ('AgnesG') to search within
 - subset: subset to return, one of shortsus/longsus/nonquote/quote/all. Default 'all' (i.e. all text)
 - contextsize: Size of context window around subset. Default 0.
+- metadata: Optional data to return, currently only 'book_titles' to return full book titles
 
 Parameters should be provided in querystring format, for example::
 
@@ -70,13 +71,15 @@ Examples:
     ], "version":{"corpora":"master:fc4de7c", "clic":"1.6:95bf699"}}
 
 """
-def subset(cdb, corpora=['dickens'], subset=['all'], contextsize=['0']):
+def subset(cdb, corpora=['dickens'], subset=['all'], contextsize=['0'], metadata=[]):
     """
     Main entry function for subset search
 
     - corpora: List of corpora / book names
     - subset: Subset(s) to search for.
     - contextsize: Size of context window, defaults to none.
+    - metadata, Array of extra metadata to provide with result, some of
+      - 'book_titles' (return dict of book IDs to titles at end of result)
     """
     contextsize = int(contextsize[0])
 
@@ -103,14 +106,20 @@ def subset(cdb, corpora=['dickens'], subset=['all'], contextsize=['0']):
 
     yield {} # Return empty header
     cur_chapter = None
+    book_ids = set()
     for (chapter_id, offset_start, offset_end) in cdb.rdb_query(query, params):
         if not cur_chapter or cur_chapter != chapter_id:
             cur_chapter = chapter_id
             ch = cdb.get_chapter(cur_chapter)
             (book_id, chapter_num, count_prev_chap, total_word) = cdb.get_chapter_word_counts(chapter_id)
+            book_ids.add(book_id)
 
         (_, para_chap, sent_chap) = cdb.get_word(chapter_id, [0, offset_start])
         yield ch.get_conc_line(offset_start, offset_end - offset_start, contextsize) + [
                 [book_id, chapter_num, para_chap, sent_chap],
                 [count_prev_chap + int(offset_start), total_word, chapter_id, offset_start, offset_end],
         ]
+    if 'book_titles' in metadata:
+        yield ('footer', dict(
+            book_titles=cdb.get_book_titles(book_ids),
+        ))